+++ scripts/ui/pds_ui.py
@@ -13,7 +13,9 @@ from PyQt5 import QtCore, QtWidgets
 
 
 class Pds_Ui(QtWidgets.QWidget):
-    def __init__(self, width: float, height: float, publisher, parent=None, MainWindow=None):
+    def __init__(
+        self, width: float, height: float, publisher, parent=None, MainWindow=None
+    ):
         super().__init__()
         self.width = width
         self.height = height
@@ -24,7 +26,9 @@ class Pds_Ui(QtWidgets.QWidget):
         self.setupUi(MainWindow)
 
     def setupUi(self, MainWindow):
-        self.log_browser = Log_browser(self.width, self.height, self.publisher, self.parent)
+        self.log_browser = Log_browser(
+            self.width, self.height, self.publisher, self.parent
+        )
         self.log_browser.setup()
 
         self.wheel_table = Wheel_table(
@@ -41,26 +45,28 @@ class Pds_Ui(QtWidgets.QWidget):
         )
         self.arm_table.setup()
 
-        self.pds_table = Pds_table(self.width, self.height, self.parent, self.width / 1.45, self.height / 2.7)
+        self.pds_table = Pds_table(
+            self.width, self.height, self.parent, self.width / 1.45, self.height / 2.7
+        )
         self.pds_table.setup()
 
         self.list_commands_button = QtWidgets.QPushButton(self.parent)
         self.list_commands_button.setGeometry(
             QtCore.QRect(
-                self.width / 2,
-                self.height / 5.14,
-                self.width / 14.66,
-                self.height / 15.21,
+                int(self.width / 2),
+                int(self.height / 5.14),
+                int(self.width / 14.66),
+                int(self.height / 15.21),
             )
         )
         self.list_commands_button.setObjectName("list_commands_button")
         self.stop_button = QtWidgets.QPushButton(self.parent)
         self.stop_button.setGeometry(
             QtCore.QRect(
-                self.width / 3.49,
-                self.height / 3.72,
-                self.width / 14.66,
-                self.height / 15.21,
+                int(self.width / 3.49),
+                int(self.height / 3.72),
+                int(self.width / 14.66),
+                int(self.height / 15.21),
             )
         )
         self.stop_button.setStyleSheet("background-color: red")
@@ -68,35 +74,40 @@ class Pds_Ui(QtWidgets.QWidget):
         self.reset_general_flags_button = QtWidgets.QPushButton(self.parent)
         self.reset_general_flags_button.setGeometry(
             QtCore.QRect(
-                self.width / 2.53,
-                self.height / 5.14,
-                self.width / 14.66,
-                self.height / 15.21,
+                int(self.width / 2.53),
+                int(self.height / 5.14),
+                int(self.width / 14.66),
+                int(self.height / 15.21),
             )
         )
         self.enable_motors_button = QtWidgets.QPushButton(self.parent)
         self.enable_motors_button.setGeometry(
-            QtCore.QRect(self.width / 3.49, self.height / 5.14, self.width / 14.66, self.height / 15.21)
+            QtCore.QRect(
+                int(self.width / 3.49),
+                int(self.height / 5.14),
+                int(self.width / 14.66),
+                int(self.height / 15.21),
+            )
         )
         self.enable_motors_button.setObjectName("enable_motors_button")
         self.reset_general_flags_button.setObjectName("reset_general_flags_button")
         self.reset_current_flags_button = QtWidgets.QPushButton(self.parent)
         self.reset_current_flags_button.setGeometry(
             QtCore.QRect(
-                self.width / 2.53,
-                self.height / 3.72,
-                self.width / 14.66,
-                self.height / 15.21,
+                int(self.width / 2.53),
+                int(self.height / 3.72),
+                int(self.width / 14.66),
+                int(self.height / 15.21),
             )
         )
         self.reset_current_flags_button.setObjectName("reset_current_flags_button")
         self.layoutWidget1 = QtWidgets.QWidget(self.parent)
         self.layoutWidget1.setGeometry(
             QtCore.QRect(
-                self.width / 2.53,
-                self.height / 21.6,
-                self.width / 5.98,
-                self.height / 12,
+                int(self.width / 2.53),
+                int(self.height / 21.6),
+                int(self.width / 5.98),
+                int(self.height / 12),
             )
         )
         self.layoutWidget1.setObjectName("layoutWidget1")
@@ -133,10 +144,10 @@ class Pds_Ui(QtWidgets.QWidget):
         self.layoutWidget2 = QtWidgets.QWidget(self.parent)
         self.layoutWidget2.setGeometry(
             QtCore.QRect(
-                self.width / 3.49,
-                self.height / 21.6,
-                self.width / 12.39,
-                self.height / 10.69,
+                int(self.width / 3.49),
+                int(self.height / 21.6),
+                int(self.width / 12.39),
+                int(self.height / 10.69),
             )
         )
         self.layoutWidget2.setObjectName("layoutWidget2")
@@ -153,10 +164,10 @@ class Pds_Ui(QtWidgets.QWidget):
         self.layoutWidget3 = QtWidgets.QWidget(self.parent)
         self.layoutWidget3.setGeometry(
             QtCore.QRect(
-                self.width / 1.7,
-                self.height / 20,
-                self.width / 19.2,
-                self.height / 21.6,
+                int(self.width / 1.7),
+                int(self.height / 20),
+                int(self.width / 19.2),
+                int(self.height / 21.6),
             )
         )
         self.layoutWidget3.setObjectName("layoutWidget3")
@@ -171,14 +182,16 @@ class Pds_Ui(QtWidgets.QWidget):
         self.fan1_speed_input.setDecimals(1)
         self.fan1_speed_input.setProperty("value", 100.0)
         self.fan1_speed_input.setObjectName("fan1_speed_input")
-        self.formLayout.setWidget(1, QtWidgets.QFormLayout.LabelRole, self.fan1_speed_input)
+        self.formLayout.setWidget(
+            1, QtWidgets.QFormLayout.LabelRole, self.fan1_speed_input
+        )
         self.layoutWidget4 = QtWidgets.QWidget(self.parent)
         self.layoutWidget4.setGeometry(
             QtCore.QRect(
-                self.width / 1.7,
-                self.height / 9,
-                self.width / 19.2,
-                self.height / 19.29,
+                int(self.width / 1.7),
+                int(self.height / 9),
+                int(self.width / 19.2),
+                int(self.height / 19.29),
             )
         )
         self.layoutWidget4.setObjectName("layoutWidget4")
@@ -190,7 +203,9 @@ class Pds_Ui(QtWidgets.QWidget):
         self.fan2_speed_input.setDecimals(1)
         self.fan2_speed_input.setProperty("value", 100.0)
         self.fan2_speed_input.setObjectName("fan2_speed_input")
-        self.formLayout_2.setWidget(1, QtWidgets.QFormLayout.LabelRole, self.fan2_speed_input)
+        self.formLayout_2.setWidget(
+            1, QtWidgets.QFormLayout.LabelRole, self.fan2_speed_input
+        )
         self.label = QtWidgets.QLabel(self.layoutWidget4)
         self.label.setObjectName("label")
         self.formLayout_2.setWidget(0, QtWidgets.QFormLayout.LabelRole, self.label)
@@ -205,18 +220,28 @@ class Pds_Ui(QtWidgets.QWidget):
 
     def retranslateUi(self):
         _translate = QtCore.QCoreApplication.translate
-        self.list_commands_button.setText(_translate("MainWindow", "List Commands\n" "(L)"))
+        self.list_commands_button.setText(
+            _translate("MainWindow", "List Commands\n" "(L)")
+        )
         self.stop_button.setText(_translate("MainWindow", "STOP\n(Space)"))
-        self.enable_motors_button.setText(_translate("MainWindow", "Enable all Motors\n(Ctrl+Q)"))
-        self.reset_general_flags_button.setText(_translate("MainWindow", "Reset General\n" "Flags"))
-        self.reset_current_flags_button.setText(_translate("MainWindow", "Reset Current\n" "Flags"))
+        self.enable_motors_button.setText(
+            _translate("MainWindow", "Enable all Motors\n(Ctrl+Q)")
+        )
+        self.reset_general_flags_button.setText(
+            _translate("MainWindow", "Reset General\n" "Flags")
+        )
+        self.reset_current_flags_button.setText(
+            _translate("MainWindow", "Reset Current\n" "Flags")
+        )
         self.motor1.setText(_translate("MainWindow", "Motor 1"))
         self.motor2.setText(_translate("MainWindow", "Motor 2"))
         self.motor3.setText(_translate("MainWindow", "Motor 3"))
         self.motor4.setText(_translate("MainWindow", "Motor 4"))
         self.motor5.setText(_translate("MainWindow", "Motor 5"))
         self.motor6.setText(_translate("MainWindow", "Motor 6"))
-        self.command_listener_checkbox.setText(_translate("MainWindow", "Command Listener"))
+        self.command_listener_checkbox.setText(
+            _translate("MainWindow", "Command Listener")
+        )
         self.auto_mode_checkbox.setText(_translate("MainWindow", "Auto Mode"))
         self.label_2.setText(_translate("MainWindow", "Fan 1 Speed"))
         self.label.setText(_translate("MainWindow", "Fan 2 Speed"))
